point level 합체
